requirements.txt
filterpy==1.4.5
scikit-image==0.17.2
lap==0.4.0
ffmpeg-python
seaborn
Pillow
numpy